Add initialEmail parameter and update onSignInComplete
Added an optional initialEmail to pass in a saved email address so that the email field is already populated.

Passed in the email to the onSignInComplete() callback, so that the email can be used inside (eg. for saving to sharedPreferences)

Also, added a focus node for the password field to move the focus to it if the initialEmail is non-empty.